requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
google-search-results>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
